Add spec for behavior that was missed
The EventModeling book expresses GWT tests only focusing on the event stream, which leads to imbalanced specs that may not focus on ensure the aggregate state is correct. In the following commits I'll refactor the specs to make sure that any command and resulting events are tested both as stream and aggregate state changes.